fix(persistence): share retention delete calculation
